feat

+++ database/migrations/2026_01_16_183250_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->text('description')->nullable();
             $table->string('image')->nullable();
-            $table->decimal('price', 10, 2);
+            $table->integer('price'); // Preço em centavos (12390 = R$ 123,90)
             $table->timestamps();
         });
     }
